(refactor) Modularized user details into its own component.
